test: e2e-identity: add an rstest fixture, test_env
The fixture is used to set up the test environment. Specifically,
it assumes that test-wire-server is already running and listening
on address given in the TEST_WIRE_SERVER_ADDR environment variable.

The fixture makes sure to start Keycloak only once by trying to
create a file named $NEXTEST_RUN_ID (this ensures uniqueness per
test run).

If the file does not exist, that means we're the first to come and
we need to start Keycloak. Once we start it, we serialize Keycloak
configuration to the file and return the configured test environment.

If the file already exists, that means that some other test came first
and we just need to read the Keycloak configuration and return the
test environment. Note that it may happen that the other test created
the file (i.e. its directory entry), but did not yet manage to
completely write serialized configuration. Because of that, we have a
loop that repeatedly tries to read from the file and deserialize
configuration from it.

Finally, note that the fixture does not take care of ACME setup yet.
That is left for some other day.