fix: be more strict when comparing if two teams are equal.
Previously two teams were considered equal if one of them were None, this
is no longer the case. This mainly affects the the generation of the dpop
access token which is a function which executes on the wire-server. The
effect is that it's no longer allowed for a client which belongs to a team
to omit the team parameter when enrolling into e2ei.