test: e2e-identity: support fetching the ID token from Authelia
This is pretty much the same as what we do with Keycloak, except for the
way we obtain an authenticated session.

With Keycloak we scrape the login form and imitate user action on the
form in order to get the access token. But there's really no need for
that complexity, so with Authelia we just

 - create a new authenticated session by providing the username and password
 - get the Authelia session cookie
 - use the session cookie when doing the authorization request

It would be nice to do the same with Keycloak eventually.

Also, the specifics of how to get an authenticated session with an IdP
really belong in IdP-specific modules, not in utils/fmk.rs. But that is
for another day.